% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRISK3_2017_test.R
\docType{data}
\name{QRISK3_2017_test}
\alias{QRISK3_2017_test}
\title{Test data for QRISK3 2017 algorithm - 2017 data}
\format{
An object of class \code{data.frame} with 48 rows and 27 columns.
}
\usage{
data(QRISK3_2017_test)
}
\description{
Data from QRISK3 original algorithm (C code) in 2017 <https://qrisk.org/three/src.php>.
The aim is to compare whether this package calculates the same score as the original algorithm.
"QRISK_C_algorithm_score" in dataset is the score calculated using original algorithm in 2017. 
It should give the same score as this package.
}
\examples{
data(QRISK3_2017_test)
str(QRISK3_2017_test)
}
\keyword{datasets}
