% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRISK3_2019_test.R
\docType{data}
\name{QRISK3_2019_test}
\alias{QRISK3_2019_test}
\title{Test data for QRISK3 2017 algorithm - 2019 data}
\format{
An object of class \code{data.frame} with 49 rows and 27 columns.
}
\usage{
data(QRISK3_2019_test)
}
\description{
Data from QRISK3 original algorithm (C code) in 2019 <https://qrisk.org/three/src.php>.
The aim is to compare whether this package calculates the same score as the original algorithm.
"QRISK_C_algorithm_score" in dataset is the score calculated using original algorithm in 2019. 
It should give the same score as this package.
This data was similar to QRISK3_2017_test except that several test values have been changed.
}
\examples{
data(QRISK3_2019_test)
str(QRISK3_2019_test)
}
\keyword{datasets}
